add file check

+++ profile.d/ssh_backup.sh
@@ -3,8 +3,14 @@
 BACKUP_LOCATION=~/Documents/ssh_key_backup_encrypted.tar.gz
 PREVIOUS_BACKUP="${BACKUP_LOCATION}.old"
 
+KEY_LOCATION=/opt/secure/ssh_key_tar.password
 
 backup_ssh(){
+    if ! [[ -f $KEY_LOCATION ]]; then
+        err password file not found
+        return 1
+    fi
+
     if [[ -f "$BACKUP_LOCATION" ]]; then
         rm -f "$PREVIOUS_BACKUP" || true
         mv "$BACKUP_LOCATION" "$PREVIOUS_BACKUP"
@@ -12,10 +18,14 @@ backup_ssh(){
 
     cd ~/.ssh || return 1
     tar --exclude=./cm -czvf - ./* | openssl enc -e -aes256 \
-        -k /opt/secure/ssh_key_tar.password -out "$BACKUP_LOCATION"
+        -k "$KEY_LOCATION" -out "$BACKUP_LOCATION"
 }
 
 restore_ssh(){
-    openssl enc -d -aes256 -k /opt/secure/ssh_key_tar.password \
+    if ! [[ -f $KEY_LOCATION ]]; then
+        err password file not found
+        return 1
+    fi
+    openssl enc -d -aes256 -k "$KEY_LOCATION" \
         -in "$BACKUP_LOCATION" | tar xvz -C ~/.ssh
 }
